moved files around. Components should go in a components folder and event-utils was unused
