+++ packages/frontend/app/(dashboard)/course/[cid]/schedule/components/CreateEventModal.tsx
