+++ packages/frontend/app/(dashboard)/course/[cid]/schedule/components/EditEventModal.tsx
