+++ packages/frontend/app/(dashboard)/course/[cid]/schedule/components/StudentSchedulePanel.tsx
