+++ packages/frontend/app/(dashboard)/course/[cid]/schedule/components/TASchedulePanel.tsx
