--- packages/frontend/app/(dashboard)/course/[cid]/schedule/event-utils.tsx
@@ -1,20 +1,0 @@
-let eventGuid = 0
-const todayStr = new Date().toISOString().replace(/T.*$/, '') // YYYY-MM-DD of today
-
-export const INITIAL_EVENTS = [
-  {
-    id: createEventId(),
-    title: 'All-day event',
-    daysOfWeek: ['4'],
-    start: todayStr,
-  },
-  {
-    id: createEventId(),
-    title: 'Timed event',
-    start: todayStr + 'T12:00:00',
-  },
-]
-
-export function createEventId() {
-  return String(eventGuid++)
-}
